+++ src/inmanta/server/__init__.py
@@ -21,3 +21,5 @@
 SLICE_SERVER = "server"
 SLICE_AGENT_MANAGER = "agentmanager"
 SLICE_SESSION_MANAGER = "session"
+SLICE_DATABASE = "core.database"
+SLICE_TRANSPORT = "core.transport"