+++ src/inmanta/server/protocol.py
@@ -15,8 +15,6 @@
 
     Contact: code@inmanta.com
 """
-from _asyncio import Task
-
 import inmanta.protocol.endpoints
 from inmanta.types import JsonType
 from inmanta.util import Scheduler, TaskHandler, stable_depth_first, CycleException
@@ -24,8 +22,8 @@ from inmanta.protocol import Client, handle, methods
 from inmanta.protocol import common, endpoints
 from inmanta.protocol.rest import server
 
-from inmanta import config as inmanta_config, data
-from inmanta.server import config as opt, SLICE_SESSION_MANAGER
+from inmanta import config as inmanta_config
+from inmanta.server import config as opt, SLICE_SESSION_MANAGER, SLICE_TRANSPORT
 
 from tornado import gen, queues, web, routing
 from tornado.ioloop import IOLoop
@@ -36,8 +34,7 @@ import logging
 import asyncio
 import time
 import uuid
-import abc
-
+from collections import defaultdict
 
 LOGGER = logging.getLogger(__name__)
 
@@ -90,6 +87,7 @@ class Server(endpoints.Endpoint):
         self.add_slice(self.sessions_handler)
 
         self._transport = server.RESTServer(self.sessions_handler, self.id)
+        self.add_slice(TransportSlice(self))
         self.running = False
 
     def add_slice(self, slice: "ServerSlice") -> None:
@@ -114,10 +112,16 @@ class Server(endpoints.Endpoint):
     id = property(get_id)
 
     def _order_slices(self) -> List["ServerSlice"]:
-        edges: Dict[str, List[str]] = {myslice.name: myslice.get_dependencies() for myslice in self.get_slices().values()}
+        edges: Dict[str, Set[str]] = defaultdict(set)
+
+        for slice in self.get_slices().values():
+            edges[slice.name].update(slice.get_dependencies())
+            for depby in slice.get_dependened_by():
+                edges[depby].add(slice.name)
+
         names = list(edges.keys())
         try:
-            order = stable_depth_first(names, edges)
+            order = stable_depth_first(names, {k: list(v) for k, v in edges.items()})
         except CycleException as e:
             raise ServerStartFailure("Dependency cycle between server slices " + ",".join(e.nodes)) from e
 
@@ -148,22 +152,20 @@ class Server(endpoints.Endpoint):
         LOGGER.debug("Starting Server Rest Endpoint")
         self.running = True
 
-        await self.connect_database()
-
         for my_slice in self._get_slice_sequence():
             try:
+                LOGGER.debug("Pre Starting %s", my_slice.name)
                 await my_slice.prestart(self)
             except Exception as e:
                 raise SliceStartupException(my_slice.name, e)
 
         for my_slice in self._get_slice_sequence():
             try:
+                LOGGER.debug("Starting %s", my_slice.name)
                 await my_slice.start()
                 self._handlers.extend(my_slice.get_handlers())
             except Exception as e:
                 raise SliceStartupException(my_slice.name, e)
-
-        await self._transport.start(self.get_slices().values(), self._handlers)
 
     async def stop(self) -> None:
         """
@@ -173,40 +175,40 @@ class Server(endpoints.Endpoint):
             This prevents database connection from being closed too early. This order in which the endpoint
             are started, is hardcoded in the get_server_slices() method in server/bootloader.py
         """
-        await super(Server, self).stop()
         if not self.running:
             return
-
         self.running = False
-        LOGGER.debug("Stopping Server Rest Endpoint")
-        await self._transport.stop()
-        for endpoint in reversed(list(self._get_slice_sequence())):
-            LOGGER.debug("Stopping %s", endpoint)
+
+        await super(Server, self).stop()
+
+        order = list(reversed(self._get_slice_sequence()))
+
+        for endpoint in order:
+            LOGGER.debug("Pre Stopping %s", endpoint.name)
+            await endpoint.prestop()
+
+        for endpoint in order:
+            LOGGER.debug("Stopping %s", endpoint.name)
             await endpoint.stop()
-
-        await self._transport.join()
-        await self.disconnect_database()
-
-    async def connect_database(self) -> None:
-        """ Connect to the database
-        """
-        database_host = opt.db_host.get()
-        database_port = opt.db_port.get()
-
-        database_username = opt.db_username.get()
-        database_password = opt.db_password.get()
-        await data.connect(database_host, database_port, opt.db_name.get(), database_username, database_password)
-        LOGGER.info("Connected to PostgreSQL database %s on %s:%d", opt.db_name.get(), database_host, database_port)
-
-    async def disconnect_database(self) -> None:
-        """ Disconnect the database
-        """
-        await data.disconnect()
 
 
 class ServerSlice(inmanta.protocol.endpoints.CallTarget, TaskHandler):
     """
-        An API serving part of the server.
+        Base class for server extensions offering zero or more api endpoints
+
+        Extensions developers should override the lifecycle methods:
+
+        * :func:`ServerSlice.prestart`
+        * :func:`ServerSlice.start`
+        * :func:`ServerSlice.prestop`
+        * :func:`ServerSlice.stop`
+        * :func:`ServerSlice.get_dependencies`
+
+        To register endpoints that server static content, either use :func:'add_static_handler' or :func:'add_static_content'
+        To create endpoints, use the annotation based mechanism
+
+        To schedule recurring tasks, use :func:`schedule` or `self._sched`
+        To schedule background tasks, use :func:`add_background_task`
     """
 
     def __init__(self, name: str) -> None:
@@ -214,34 +216,75 @@ class ServerSlice(inmanta.protocol.endpoints.CallTarget, TaskHandler):
 
         self._name: str = name
         self._handlers: List[routing.Rule] = []
-        self._sched = Scheduler("server slice")  # FIXME: why has each slice its own scheduler?
-        self.running: bool = False  # for debugging
-        self._background_tasks: Set[Task] = set()
+        self._sched = Scheduler(f"server slice {name}")
+        # is shutdown in progress?
+        self._stopping: bool = False
 
-    @abc.abstractmethod
+    def is_stopping(self):
+        """True when prestop has been called."""
+        return self._stopping
+
     async def prestart(self, server: Server) -> None:
-        """Called by the RestServer host prior to start, can be used to collect references to other server slices"""
+        """
+        Called by the RestServer host prior to start, can be used to collect references to other server slices
+        Dependencies are not up yet.
+        """
+        pass
 
-    @abc.abstractmethod
     async def start(self) -> None:
         """
             Start the server slice.
+
+            This method `blocks` until the slice is ready to receive calls
+
+            Dependencies are up (if present) prior to invocation of this call
         """
-        self.running = True
+        pass
+
+    async def prestop(self) -> None:
+        """
+            Always called before stop
+
+            Stop producing new work:
+            - stop timers
+            - stop listeners
+            - notify shutdown to systems depending on us (like agents)
+
+            sets is_stopping to true
+
+            But remain functional
+
+            All dependencies are up (if present)
+        """
+        self._stopping = True
+        self._sched.stop()
 
     async def stop(self) -> None:
-        self.running = False
-        self._sched.stop()
+        """
+            Go down
+
+            All dependencies are up (if present)
+
+            This method `blocks` until the slice is down
+        """
         await super(ServerSlice, self).stop()
 
-    name = property(lambda self: self._name)
-
     def get_dependencies(self) -> List[str]:
+        """List of names of slices that must be started before this one."""
         return []
 
+    def get_dependened_by(self) -> List[str]:
+        """List of names of slices that must be started after this one."""
+        return []
+
+    # internal API towards extension framework
+    name = property(lambda self: self._name)
+
     def get_handlers(self) -> List[routing.Rule]:
+        """Get the list of """
         return self._handlers
 
+    # utility methods for extensions developers
     def schedule(self, call: Callable, interval: int = 60) -> None:
         self._sched.add_action(call, interval)
 
@@ -323,6 +366,8 @@ class Session(object):
     id = property(get_id)
 
     def expire(self, timeout: float) -> None:
+        if self.expired:
+            return
         self.expired = True
         if self._callhandle is not None:
             IOLoop.current().remove_timeout(self._callhandle)
@@ -410,6 +455,31 @@ class SessionListener(object):
 
 
 # Internals
+class TransportSlice(ServerSlice):
+    """Slice to manage the listening socket"""
+
+    def __init__(self, server: Server):
+        super(TransportSlice, self).__init__(SLICE_TRANSPORT)
+        self.server = server
+
+    def get_dependencies(self) -> List[str]:
+        """All Slices with an http endpoint should depend on this one using :func:`get_dependened_by`"""
+        return []
+
+    async def start(self) -> None:
+        await super(TransportSlice, self).start()
+        await self.server._transport.start(self.server.get_slices().values(), self.server._handlers)
+
+    async def prestop(self) -> None:
+        await super(TransportSlice, self).prestop()
+        LOGGER.debug("Stopping Server Rest Endpoint")
+        await self.server._transport.stop()
+
+    async def stop(self) -> None:
+        await super(TransportSlice, self).stop()
+        await self.server._transport.join()
+
+
 class SessionManager(ServerSlice):
     """
         A service that receives method calls over one or more transports
@@ -439,24 +509,15 @@ class SessionManager(ServerSlice):
     def add_listener(self, listener: SessionListener) -> None:
         self.listeners.append(listener)
 
-    async def prestart(self, server: Server) -> None:
-        """Called by the RestServer host prior to start, can be used to collect references to other server slices"""
-
-    async def start(self) -> None:
-        """
-            Start the server slice.
-        """
-        await super().start()
-
-    async def stop(self) -> None:
-        """
-            Stop the end-point and all of its transports
-        """
-        await super().stop()
+    async def prestop(self) -> None:
+        await super(SessionManager, self).prestop()
         # terminate all sessions cleanly
         for session in self._sessions.copy().values():
             session.expire(0)
             session.abort()
+
+    def get_dependened_by(self) -> List[str]:
+        return [SLICE_TRANSPORT]
 
     def validate_sid(self, sid: uuid.UUID) -> bool:
         if isinstance(sid, str):
@@ -486,9 +547,9 @@ class SessionManager(ServerSlice):
 
     def expire(self, session: Session, timeout: float) -> None:
         LOGGER.debug("Expired session with id %s, last seen %d seconds ago" % (session.get_id(), timeout))
+        del self._sessions[session.id]
         for listener in self.listeners:
             listener.expire(session, timeout)
-        del self._sessions[session.id]
 
     def seen(self, session: Session, endpoint_names: List[str]) -> None:
         LOGGER.debug("Seen session with id %s" % (session.get_id()))