+++ tests/test_agent_manager.py
@@ -83,7 +83,6 @@ async def test_primary_selection(init_dataclasses_and_load_schema):
     server.add_background_task.side_effect = futures
     am = AgentManager(server, False)
     am.add_background_task = futures
-    am.add_future = futures
     am.running = True
 
     async def assert_agent(name: str, state: str, sid: UUID):
@@ -326,7 +325,6 @@ async def test_db_clean(init_dataclasses_and_load_schema):
     server.add_background_task.side_effect = futures
     am = AgentManager(server, False)
     am.add_background_task = futures
-    am.add_future = futures
     am.running = True
 
     async def assert_agent(name: str, state: str, sid: UUID):
@@ -385,9 +383,8 @@ async def test_db_clean(init_dataclasses_and_load_schema):
     # failover
     am = AgentManager(server, False)
     am.add_background_task = futures
-    am.add_future = futures
     am.running = True
-    await am.clean_db()
+    await am._clean_db()
 
     # one session
     ts1 = MockSession(uuid4(), env.id, ["agent1", "agent2"], "ts1")